test: Isolate SystemMessageId test to EmbeddedConnector
SystemMessageId test doesn't work with websocket/ipc since those reset
the message Id.
